Phones
Create the phones enum, to represent Klatt phones, and begin populating data. As of this commit, the parameter targets for all phones are done.

In addition, hook up noise sources for synthesis.

However, there are still gaps for things set by rules and not parameter targets. That includes stops (frication is zero in the targets, but set by rule), nasals, and quasi-sinusoidal voicing.